Clean up actions
In general, don't create an action for something that is only used server-side. It's unnecessary and redundant. Actions should be used primarily when we have calls that need to be made client-side, or where otherwise mandated, like in the case of setting cookies (since this needs to happen in the browser).