Refactor and restructure application dashboard (#18)
* Run prettier

* Clean up actions

In general, don't create an action for something that is only used server-side. It's unnecessary and redundant. Actions should be used primarily when we have calls that need to be made client-side, or where otherwise mandated, like in the case of setting cookies (since this needs to happen in the browser).

* Make the partner select actually reflect the selected partner

The partner select is not getting the value of the set cookie. This means when you reload, the select resets even though a partner is still selected in the cookie. Also, there was no way to clear the selection.

* Clean up some data types and prepare statements

* Clean up report and reportItem data queries and prepare statements

* Clean up reportItems

Remove unnecessary actions, cache queries, consolidate into upsert

* Clean up actions, query and search params for reports

* Refactor partners query and search params

* Refactor reports queries and search params

* Update eslint config and cleanup generally

* Add current value to select and line it up with buttons