Remove 'server-only' from unneeded files (#24)
As it turns out, adding 'server-only' to files that don't actually need or care about it, breaks stuff. In this case, adding it to schema and helper files that drizzle-kit was trying to read was breaking it. There's nothing sensitive in these files, nor would they ever be included in our client-side package to begin with, so we should exclude the directive.